3.0 - The finished product!
I've redesigned the main menu, added sounds to the game, and cleaned up
the code to reduce the amount of global variables. This is the end of
the month and I mark this release as the final version. If I come back
to this, it will be for minor changes but I won't for a while.